Address feedback and refactor
