+++ tools/mirror-report/package.json
@@ -8,10 +8,12 @@
   },
   "dependencies": {
     "commander": "^12.1.0",
+    "json-bigint": "^1.0.0",
     "node-fetch": "^3.3.2"
   },
   "description": "Generates a report showing an account's hbar transfers",
   "devDependencies": {
+    "@fetch-mock/jest": "^0.2.7",
     "jest": "^29.7.0"
   },
   "homepage": "https://github.com/hashgraph/hedera-mirror-node/tree/main/tools/mirror-report",
@@ -29,7 +31,7 @@
     "url": "git+https://github.com/hashgraph/hedera-mirror-node.git"
   },
   "scripts": {
-    "start": "node ./index.js",
+    "start": "node --experimental-specifier-resolution=node ./index.js",
     "lint": "eslint --ignore-pattern node_modules/ --fix .",
     "test": "node --experimental-specifier-resolution=node --experimental-vm-modules node_modules/jest/bin/jest.js"
   },
